args in testAction, payload as array, check for no dispatches
I reused the testAction for all my action tests, so I made it a little more robust so I can pass in mocked arguments into the action as an array, and also check for more than one item in the payload. Lastly, I added a check at the end to handle scenarios where an action is called that leads to no mutations being dispatched.